+++ xmake/core/package/package.lua
@@ -737,9 +737,9 @@ function _instance:script(name, generic)
         -- `@linux|x86_64`
         -- `@macosx,linux`
         -- `android@macosx,linux`
-        -- `android|armv7-a@macosx,linux`
-        -- `android|armv7-a@macosx,linux|x86_64`
-        -- `android|armv7-a@linux|x86_64`
+        -- `android|armeabi-v7a@macosx,linux`
+        -- `android|armeabi-v7a@macosx,linux|x86_64`
+        -- `android|armeabi-v7a@linux|x86_64`
         --
         for _pattern, _script in pairs(script) do
             local hosts = {}
@@ -806,6 +806,11 @@ function _instance:fetch(opt)
     -- true: only find system package
     -- false: only find xmake packages
     local system = opt.system or self:requireinfo().system
+    if self:is3rd() then
+        -- we need ignore `{system = true/false}` argument if be 3rd package
+        -- @see https://github.com/xmake-io/xmake/issues/726
+        system = nil
+    end
 
     -- fetch binary tool?
     fetchinfo = nil