+++ xmake/modules/private/action/trybuild/autotools.lua
@@ -51,11 +51,20 @@ end
 function _get_buildenvs()
     local envs = {}
     if is_plat(os.subhost()) then
-        local cflags   = table.join(table.wrap(config.get("cxflags")), config.get("cflags"))
-        local cxxflags = table.join(table.wrap(config.get("cxflags")), config.get("cxxflags"))
+        local cflags   = table.join(table.wrap(package:config("cxflags")), package:config("cflags"))
+        local cxxflags = table.join(table.wrap(package:config("cxflags")), package:config("cxxflags"))
+        local asflags  = table.copy(table.wrap(package:config("asflags")))
+        local ldflags  = table.copy(table.wrap(package:config("ldflags")))
+        if package:is_plat("linux") and package:is_arch("i386") then
+            table.insert(cflags,   "-m32")
+            table.insert(cxxflags, "-m32")
+            table.insert(asflags,  "-m32")
+            table.insert(ldflags,  "-m32")
+        end
         envs.CFLAGS    = table.concat(cflags, ' ')
         envs.CXXFLAGS  = table.concat(cxxflags, ' ')
-        envs.ASFLAGS   = table.concat(table.wrap(config.get("asflags")), ' ')
+        envs.ASFLAGS   = table.concat(asflags, ' ')
+        envs.LDFLAGS   = table.concat(ldflags, ' ')
     else
         local cflags   = table.join(table.wrap(_get_buildenv("cxflags")), _get_buildenv("cflags"))
         local cxxflags = table.join(table.wrap(_get_buildenv("cxflags")), _get_buildenv("cxxflags"))
@@ -129,15 +138,18 @@ function _get_configs(artifacts_dir)
             -- @see https://developer.android.com/ndk/guides/other_build_systems#autoconf
             local triples = 
             {
-                ["armv5te"]     = "arm-linux-androideabi",
-                ["armv7-a"]     = "arm-linux-androideabi",
+                ["armv5te"]     = "arm-linux-androideabi",  -- deprecated
+                ["armv7-a"]     = "arm-linux-androideabi",  -- deprecated
+                ["armeabi"]     = "arm-linux-androideabi",  -- removed in ndk r17
+                ["armeabi-v7a"] = "arm-linux-androideabi",
                 ["arm64-v8a"]   = "aarch64-linux-android",
-                i386            = "i686-linux-android",
+                i386            = "i686-linux-android",     -- deprecated
+                x86             = "i686-linux-android",
                 x86_64          = "x86_64-linux-android",
-                mips            = "mips-linux-android",
-                mips64          = "mips64-linux-android"
+                mips            = "mips-linux-android",     -- removed in ndk r17
+                mips64          = "mips64-linux-android"    -- removed in ndk r17
             }
-            table.insert(configs, "--host=" .. (triples[config.arch()] or triples["armv7-a"]))
+            table.insert(configs, "--host=" .. (triples[config.arch()] or triples["armeabi-v7a"]))
         elseif is_plat("mingw") then
             local triples = 
             { 
@@ -196,7 +208,11 @@ function build()
     end
 
     -- do build
-    os.vexec("make -j" .. option.get("jobs"))
+    local argv = {"-j" .. option.get("jobs")}
+    if option.get("verbose") then
+        table.insert(argv, "V=1")
+    end
+    os.vexecv("make", argv)
     os.vexec("make install")
     cprint("output to ${bright}%s", artifacts_dir)
     cprint("${color.success}build ok!")