+++ xmake/platforms/android/config.lua
@@ -74,12 +74,13 @@ function _toolchains()
     local ar         = toolchain("the static library archiver")
     local ex         = toolchain("the static library extractor")
     local ranlib     = toolchain("the static library index generator")
+    local strip      = toolchain("the symbols stripper")
     local as         = toolchain("the assember")
     local rc         = toolchain("the rust compiler")
     local rc_ld      = toolchain("the rust linker")
     local rc_sh      = toolchain("the rust shared library linker")
     local rc_ar      = toolchain("the rust static library archiver")
-    local toolchains = {cc = cc, cxx = cxx, cpp = cpp, as = as, ld = ld, sh = sh, ar = ar, ex = ex, ranlib = ranlib, 
+    local toolchains = {cc = cc, cxx = cxx, cpp = cpp, as = as, ld = ld, sh = sh, ar = ar, ex = ex, ranlib = ranlib, strip = strip, 
                         rc = rc, ["rc-ld"] = rc_ld, ["rc-sh"] = rc_sh, ["rc-ar"] = rc_ar}
 
     -- init the c compiler
@@ -113,6 +114,9 @@ function _toolchains()
     -- init the static library index generator
     ranlib:add({name = "ranlib", cross = cross, pathes = gcc_toolchain_bin}, "ranlib")
 
+    -- init the symbols stripper
+    strip:add({name = "strip", cross = cross, pathes = gcc_toolchain_bin}, "strip")
+
     -- init the rust compiler and linker
     rc:add("$(env RC)", "rustc")
     rc_ld:add("$(env RC)", "rustc")
@@ -134,7 +138,7 @@ function main(platform, name)
     else
 
         -- check arch
-        check_arch(config, "armv7-a")
+        check_arch(config, "armeabi-v7a")
 
         -- check ndk
         _check_ndk()