+++ xmake/rules/c++/xmake.lua
@@ -44,4 +44,16 @@ rule("c++.build")
 
 -- define rule: cpp
 rule("c++")
-    add_deps("c++.build", "c.build", "utils.merge.object", "utils.merge.archive")
+
+    -- add build rules
+    add_deps("c++.build", "c.build")
+
+    -- inherit links and linkdirs of all dependent targets by default
+    add_deps("utils.inherit.links")
+
+    -- support `add_files("src/*.o")` and `add_files("src/*.a")` to merge object and archive files to target
+    add_deps("utils.merge.object", "utils.merge.archive")
+
+    -- we attempt to extract symbols to the independent file and 
+    -- strip self-target binary if `set_symbols("debug")` and `set_strip("all")` are enabled
+    add_deps("utils.symbols.extract")