+++ xmake/themes/default/xmake.lua
@@ -65,10 +65,16 @@ theme("default")
     set_color("dump.string", "magenta bright")
     set_color("dump.string_quote", "magenta")
     set_color("dump.keyword", "blue")
-    set_color("dump.number", "green")
+    set_color("dump.number", "green bright")
     set_color("dump.function", "cyan")
 
     -- menu
     set_color("menu.main.task.name", "magenta")
     set_color("menu.option.name", "green")
     set_color("menu.usage", "cyan")
+
+    -- interactive mode
+    set_text("interactive.prompt", "xmake>")
+    set_text("interactive.prompt2", "xmake>>")
+    set_color("interactive.prompt", "green")
+    set_color("interactive.prompt2", "green")