+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### New features
 
+* [#727](https://github.com/xmake-io/xmake/issues/727): Strip and generate debug symbols file (.so/.dSYM) for android/ios program
+* [#687](https://github.com/xmake-io/xmake/issues/687): Support to generate objc/bundle program.
+* [#743](https://github.com/xmake-io/xmake/issues/743): Support to generate objc/framework program.
+
+## v2.3.2
+
+### New features
+
 * Add powershell theme for powershell terminal
 * Add `xmake --dry-run -v` to dry run building target and only show verbose build command.
 * [#712](https://github.com/xmake-io/xmake/issues/712): Add sdcc platform and support sdcc compiler
@@ -12,6 +20,9 @@
 
 * [#589](https://github.com/xmake-io/xmake/issues/589): Improve and optimize build speed, supports parallel compilation and linking across targets
 * Improve the ninja/cmake generator
+* [#728](https://github.com/xmake-io/xmake/issues/728): Improve os.cp to support reserve source directory structure
+* [#732](https://github.com/xmake-io/xmake/issues/732): Improve find_package to support `homebrew/cmake` pacakges
+* [#695](https://github.com/xmake-io/xmake/issues/695): Improve android abi
 
 ### Bugs fixed
 
@@ -19,6 +30,7 @@
 * [#718](https://github.com/xmake-io/xmake/issues/718): Fix download cache bug for package
 * [#722](https://github.com/xmake-io/xmake/issues/722): Fix invalid package deps
 * [#719](https://github.com/xmake-io/xmake/issues/719): Fix process exit bug
+* [#720](https://github.com/xmake-io/xmake/issues/720): Fix compile_commands generator
 
 ## v2.3.1
 
@@ -701,6 +713,14 @@
 
 ### 新特性
 
+* [#727](https://github.com/xmake-io/xmake/issues/727): 支持为android, ios程序生成.so/.dSYM符号文件
+* [#687](https://github.com/xmake-io/xmake/issues/687): 支持编译生成objc/bundle程序
+* [#743](https://github.com/xmake-io/xmake/issues/743): 支持编译生成objc/framework程序
+
+## v2.3.2
+
+### 新特性
+
 * 添加powershell色彩主题用于powershell终端下背景色显示
 * 添加`xmake --dry-run -v`命令去空运行构建，仅仅为了查看详细的构建命令
 * [#712](https://github.com/xmake-io/xmake/issues/712): 添加sdcc平台，并且支持sdcc编译器
@@ -709,6 +729,9 @@
 
 * [#589](https://github.com/xmake-io/xmake/issues/589): 改进优化构建速度，支持跨目标间并行编译和link，编译速度和ninja基本持平
 * 改进ninja/cmake工程文件生成器插件
+* [#728](https://github.com/xmake-io/xmake/issues/728): 改进os.cp支持保留源目录结构层级的递归复制
+* [#732](https://github.com/xmake-io/xmake/issues/732): 改进find_package支持查找homebrew/cmake安装的包
+* [#695](https://github.com/xmake-io/xmake/issues/695): 改进采用android ndk最新的abi命名
 
 ### Bugs修复
 
@@ -716,6 +739,7 @@
 * [#718](https://github.com/xmake-io/xmake/issues/718): 修复依赖包下载在多镜像时一定概率缓存失效问题
 * [#722](https://github.com/xmake-io/xmake/issues/722): 修复无效的包依赖导致安装死循环问题
 * [#719](https://github.com/xmake-io/xmake/issues/719): 修复windows下主进程收到ctrlc后，.bat子进程没能立即退出的问题
+* [#720](https://github.com/xmake-io/xmake/issues/720): 修复compile_commands生成器的路径转义问题
 
 ## v2.3.1